+++ src/tinyhhg_core/gridtransferoperators/P2toP2QuadraticProlongation.cpp
@@ -1,8 +1,6 @@
 #include "tinyhhg_core/gridtransferoperators/P2toP2QuadraticProlongation.hpp"
-
+#include "tinyhhg_core/gridtransferoperators/generatedKernels/all.hpp"
 #include "tinyhhg_core/FunctionMemory.hpp"
-#include "tinyhhg_core/gridtransferoperators/generatedKernels/GeneratedKernelsP2MacroCell3D.hpp"
-#include "tinyhhg_core/gridtransferoperators/generatedKernels/GeneratedKernelsP2MacroFace2D.hpp"
 #include "tinyhhg_core/p1functionspace/VertexDoFMacroCell.hpp"
 #include "tinyhhg_core/p1functionspace/VertexDoFMacroFace.hpp"
 #include "tinyhhg_core/p2functionspace/P2Function.hpp"
@@ -102,7 +100,7 @@ void P2toP2QuadraticProlongation::prolongateAdditively( const P2Function< real_t
                                                                                  &edgeFineData[firstIdxFine[eo::Y]],
                                                                                  vertexCoarseData,
                                                                                  vertexFineData,
-                                                                                 static_cast< int64_t >( coarseLevel ),
+                                                                                 static_cast< int32_t >( coarseLevel ),
                                                                                  numNeighborFacesEdge0,
                                                                                  numNeighborFacesEdge1,
                                                                                  numNeighborFacesEdge2,
@@ -117,7 +115,7 @@ void P2toP2QuadraticProlongation::prolongateAdditively( const P2Function< real_t
                                                                                &edgeFineData[firstIdxFine[eo::XY]],
                                                                                &edgeFineData[firstIdxFine[eo::Y]],
                                                                                vertexFineData,
-                                                                               static_cast< int64_t >( coarseLevel ),
+                                                                               static_cast< int32_t >( coarseLevel ),
                                                                                numNeighborFacesEdge0,
                                                                                numNeighborFacesEdge1,
                                                                                numNeighborFacesEdge2 );
@@ -223,7 +221,7 @@ void P2toP2QuadraticProlongation::prolongateAdditively3D( const P2Function< real
                                                                                &edgeFineData[firstIdxFine[eo::Z]],
                                                                                vertexCoarseData,
                                                                                vertexFineData,
-                                                                               static_cast< int64_t >( coarseLevel ),
+                                                                               static_cast< int32_t >( coarseLevel ),
                                                                                numNeighborCellsEdge0,
                                                                                numNeighborCellsEdge1,
                                                                                numNeighborCellsEdge2,
@@ -254,7 +252,7 @@ void P2toP2QuadraticProlongation::prolongateAdditively3D( const P2Function< real
                                                                              &edgeFineData[firstIdxFine[eo::YZ]],
                                                                              &edgeFineData[firstIdxFine[eo::Z]],
                                                                              vertexFineData,
-                                                                             static_cast< int64_t >( coarseLevel ),
+                                                                             static_cast< int32_t >( coarseLevel ),
                                                                              numNeighborCellsEdge0,
                                                                              numNeighborCellsEdge1,
                                                                              numNeighborCellsEdge2,
